01/21 김종록 UserDTO user_id(기본키) long 타입으로 변경, 비밀번호 찾기 에러 해결, header.jsp에 로고 경로 /spike.com 절대 경로로 변경

+++ src/main/java/com/spike/dao/UserRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.spike.dto.UserDTO;
 
-public interface UserRepository extends JpaRepository<UserDTO, Integer> {
+public interface UserRepository extends JpaRepository<UserDTO, Long> {
 
 	@Query("select s from UserDTO s where s.loginId=?1")
 	public Optional<UserDTO> findid(String id);
